WL#9323: Improved error logging: "log TNG" WL#9342: Improved error logging: log filters WL#9343: Improved error logging: log writers
- portability fixes
- valgrind fixes

+++ sql/dd/info_schema/metadata.cc
@@ -612,7 +612,7 @@ bool initialize(THD *thd)
     return true;
 
   sql_print_information("Created system views with I_S version %d",
-                        d->get_target_dd_version());
+                        (int) d->get_target_dd_version());
   return false;
 }
